feat: router and simple auth route
