+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
+    "react-router-dom": "^7.0.2",
     "subdomain_ai": "file:",
     "three": "^0.171.0"
   },
