+++ src/pages/Auth.tsx
@@ -1,0 +1,7 @@
+export const Auth = () => {
+    return(
+        <div className="bg-red">
+            Auth
+        </div>
+    )
+}
